Updated the pp_bo_form script to time based and small scale tested
Changed the while loop to work off of time and not number of triggers. Removed the alert on how many triggers there are active. Removed the actual trigger function from the form generation script. I also small scale tested it. There will need to be a larger testing effort soon.